// generated/src/app/api-models/vendor-adv-pay-account.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class VendorAdvPayAccount {
  seqVendAdvPayAcc: number;
  seqVendId: number;
  seqVendAddress: number;
  accountNumber: string;
  advPayType: string;
  replenish: string;
  companyCode: string;
  glRefCode: string;
  description: string;
  accPayType: string;
  zeroPayOption: string;
  offsetBy: string;
  offsetFlag: string;
  seqProvId: number;
  seqMembId: number;
  primarySvcDate: Date;
  claimThruDate: Date;
  userDefined1: string;
  userDefined2: string;
  userDate1: Date;
  userDate2: Date;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}